♻️(shared-presentation) move some tests in shared
